Moved UiWidgetEngine to vgc::ui (#677)
