+++ libs/vgc/ui/widget.h
@@ -19,11 +19,17 @@
 
 #include <algorithm> // std::find
 
+#include <QOpenGLBuffer>
+#include <QOpenGLFunctions_3_2_Core>
+#include <QOpenGLShaderProgram>
+#include <QOpenGLVertexArrayObject>
+
 #include <vgc/core/array.h>
 #include <vgc/core/innercore.h>
 #include <vgc/core/stringid.h>
 #include <vgc/core/vec2f.h>
 #include <vgc/graphics/engine.h>
+#include <vgc/graphics/idgenerator.h>
 #include <vgc/ui/action.h>
 #include <vgc/ui/api.h>
 #include <vgc/ui/exceptions.h>
@@ -39,6 +45,7 @@ namespace ui {
 
 VGC_DECLARE_OBJECT(Action);
 VGC_DECLARE_OBJECT(Widget);
+VGC_DECLARE_OBJECT(UiWidgetEngine);
 
 /// \typedef vgc::ui::WidgetClasses
 /// \brief Stores a set of widget classes.
@@ -725,6 +732,75 @@ private:
     friend class Style;
     Style style_;
     void onClassesChanged_();
+};
+
+/// \class vgc::widget::UiWidgetEngine
+/// \brief The graphics::Engine used by UiWidget.
+///
+/// This class is an implementation of graphics::Engine using QPainter and
+/// OpenGL calls, with the assumption to be used with a QOpenGLContext.
+///
+class VGC_UI_API UiWidgetEngine : public graphics::Engine {
+private:
+    VGC_OBJECT(UiWidgetEngine, graphics::Engine)
+
+protected:
+    /// Creates a new UiWidgetEngine. This constructor is an implementation
+    /// detail. In order to create a UiWidgetEngine, please use the following:
+    ///
+    /// \code
+    /// UiWidgetEnginePtr engine = UiWidgetEngine::create();
+    /// \endcode
+    ///
+    UiWidgetEngine();
+
+public:
+    using OpenGLFunctions = QOpenGLFunctions_3_2_Core;
+
+    /// Creates a new UiWidgetEngine.
+    ///
+    static UiWidgetEnginePtr create();
+
+    /// Initializes the engine with the given OpenGL parameters required to
+    /// perform its OpenGL calls.
+    ///
+    void initialize(OpenGLFunctions* functions,
+        QOpenGLShaderProgram* shaderProgram,
+        int posLoc, int colLoc, int projLoc, int viewLoc);
+
+    // Implementation of graphics::Engine API
+    void clear(const core::Color& color) override;
+    core::Mat4f projectionMatrix() override;
+    void setProjectionMatrix(const core::Mat4f& m) override;
+    void pushProjectionMatrix() override;
+    void popProjectionMatrix() override;
+    core::Mat4f viewMatrix() override;
+    void setViewMatrix(const core::Mat4f& m) override;
+    void pushViewMatrix() override;
+    void popViewMatrix() override;
+    Int createTriangles() override;
+    void loadTriangles(Int id, const float* data, Int length) override;
+    void drawTriangles(Int id) override;
+    void destroyTriangles(Int id) override;
+
+private:
+    OpenGLFunctions* openGLFunctions_;
+    QOpenGLShaderProgram* shaderProgram_;
+    int posLoc_;
+    int colLoc_;
+    int projLoc_;
+    int viewLoc_;
+    core::Array<core::Mat4f> projectionMatrices_;
+    core::Array<core::Mat4f> viewMatrices_;
+    graphics::IdGenerator trianglesIdGenerator_;
+    struct TrianglesBuffer {
+        // Note: we use a pointer for the VAO because copying
+        // QOpenGLVertexArrayObject is disabled
+        QOpenGLBuffer vboTriangles;
+        QOpenGLVertexArrayObject* vaoTriangles;
+        int numVertices;
+    };
+    core::Array<TrianglesBuffer> trianglesBuffers_;
 };
 
 } // namespace ui
