+++ libs/vgc/widgets/uiwidget.cpp
@@ -21,6 +21,7 @@
 
 #include <vgc/core/paths.h>
 #include <vgc/geometry/camera2d.h>
+#include <vgc/ui/widget.h>
 #include <vgc/widgets/qtutil.h>
 
 namespace vgc {
@@ -43,24 +44,12 @@ core::Mat4f toMat4f(const core::Mat4d& m) {
                (float)m(3,0), (float)m(3,1), (float)m(3,2), (float)m(3,3));
 }
 
-QMatrix4x4 toQtMatrix(const core::Mat4f& m) {
-    return QMatrix4x4(
-               m(0,0), m(0,1), m(0,2), m(0,3),
-               m(1,0), m(1,1), m(1,2), m(1,3),
-               m(2,0), m(2,1), m(2,2), m(2,3),
-               m(3,0), m(3,1), m(3,2), m(3,3));
-}
-
-struct XYRGBVertex {
-    float x, y, r, g, b;
-};
-
 } // namespace
 
 UiWidget::UiWidget(ui::WidgetPtr widget, QWidget* parent) :
     QOpenGLWidget(parent),
     widget_(widget),
-    engine_(UiWidgetEngine::create()),
+    engine_(vgc::ui::UiWidgetEngine::create()),
     isInitialized_(false)
 {
     setMouseTracking(true);
@@ -277,158 +266,6 @@ void UiWidget::onFocusRequested()
 {
     setFocus();
     // Note: Under the hood, this calls setFocus(Qt::OtherFocusReason)
-}
-
-UiWidgetEngine::UiWidgetEngine() :
-    graphics::Engine()
-{
-    projectionMatrices_.clear();
-    projectionMatrices_.append(core::Mat4f::identity);
-    viewMatrices_.clear();
-    viewMatrices_.append(core::Mat4f::identity);
-}
-
-/* static */
-UiWidgetEnginePtr UiWidgetEngine::create()
-{
-    return UiWidgetEnginePtr(new UiWidgetEngine());
-}
-
-void UiWidgetEngine::initialize(UiWidget::OpenGLFunctions* functions,
-                                QOpenGLShaderProgram* shaderProgram,
-                                int posLoc, int colLoc, int projLoc, int viewLoc)
-{
-    openGLFunctions_ = functions;
-    shaderProgram_ = shaderProgram;
-    posLoc_ = posLoc;
-    colLoc_ = colLoc;
-    projLoc_ = projLoc;
-    viewLoc_ = viewLoc;
-}
-
-void UiWidgetEngine::clear(const core::Color& color)
-{
-    openGLFunctions_->glClearColor(
-        static_cast<float>(color.r()),
-        static_cast<float>(color.g()),
-        static_cast<float>(color.b()),
-        static_cast<float>(color.a()));
-    openGLFunctions_->glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-}
-
-core::Mat4f UiWidgetEngine::projectionMatrix()
-{
-    return projectionMatrices_.last();
-}
-
-void UiWidgetEngine::setProjectionMatrix(const core::Mat4f& m)
-{
-    projectionMatrices_.last() = m;
-    shaderProgram_->setUniformValue(projLoc_, toQtMatrix(m));
-}
-
-void UiWidgetEngine::pushProjectionMatrix()
-{
-    // Note: we need a temporary copy by value, since the address
-    // of the current projection matrix might change during "append" in
-    // case of memory re-allocation.
-    core::Mat4f m = projectionMatrix();
-    projectionMatrices_.append(m);
-}
-
-void UiWidgetEngine::popProjectionMatrix()
-{
-    projectionMatrices_.removeLast();
-    shaderProgram_->setUniformValue(projLoc_, toQtMatrix(projectionMatrices_.last()));
-}
-
-core::Mat4f UiWidgetEngine::viewMatrix()
-{
-    return viewMatrices_.last();
-}
-
-void UiWidgetEngine::setViewMatrix(const core::Mat4f& m)
-{
-    viewMatrices_.last() = m;
-    shaderProgram_->setUniformValue(viewLoc_, toQtMatrix(m));
-}
-
-void UiWidgetEngine::pushViewMatrix()
-{
-    // Note: we need a temporary copy by value, since the address
-    // of the current view matrix might change during "append" in
-    // case of memory re-allocation.
-    core::Mat4f m = viewMatrix();
-    viewMatrices_.append(m);
-}
-
-void UiWidgetEngine::popViewMatrix()
-{
-    viewMatrices_.removeLast();
-    shaderProgram_->setUniformValue(viewLoc_, toQtMatrix(viewMatrices_.last()));
-}
-
-Int UiWidgetEngine::createTriangles()
-{
-    // Get or create TrianglesBuffer
-    Int id = trianglesIdGenerator_.generate();
-    if (id >= trianglesBuffers_.length()) {
-        trianglesBuffers_.append(TrianglesBuffer()); // use Array growth policy
-    }
-    if (id >= trianglesBuffers_.length()) {
-        trianglesBuffers_.resize(id+1);              // bypass Array growth policy
-    }
-    TrianglesBuffer& r = trianglesBuffers_[id];
-
-    // Create VBO/VAO for rendering triangles
-    r.vboTriangles.create();
-    r.vaoTriangles = new QOpenGLVertexArrayObject();
-    r.vaoTriangles->create();
-    GLsizei stride  = sizeof(XYRGBVertex);
-    GLvoid* posPointer = reinterpret_cast<void*>(offsetof(XYRGBVertex, x));
-    GLvoid* colPointer = reinterpret_cast<void*>(offsetof(XYRGBVertex, r));
-    GLboolean normalized = GL_FALSE;
-    r.vaoTriangles->bind();
-    r.vboTriangles.bind();
-    openGLFunctions_->glEnableVertexAttribArray(posLoc_);
-    openGLFunctions_->glEnableVertexAttribArray(colLoc_);
-    openGLFunctions_->glVertexAttribPointer(posLoc_, 2, GL_FLOAT, normalized, stride, posPointer);
-    openGLFunctions_->glVertexAttribPointer(colLoc_, 3, GL_FLOAT, normalized, stride, colPointer);
-    r.vboTriangles.release();
-    r.vaoTriangles->release();
-
-    return id;
-}
-
-void UiWidgetEngine::loadTriangles(Int id, const float* data, Int length)
-{
-    if (length < 0) {
-        throw core::NegativeIntegerError(core::format(
-            "Negative length ({}) provided to loadTriangles()", length));
-    }
-    TrianglesBuffer& r = trianglesBuffers_[id];
-    r.numVertices = length / 5;
-    r.vboTriangles.bind();
-    r.vboTriangles.allocate(data, r.numVertices * sizeof(XYRGBVertex));
-    r.vboTriangles.release();
-}
-
-void UiWidgetEngine::drawTriangles(Int id)
-{
-    TrianglesBuffer& r = trianglesBuffers_[id];
-    openGLFunctions_->glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
-    r.vaoTriangles->bind();
-    openGLFunctions_->glDrawArrays(GL_TRIANGLES, 0, r.numVertices);
-    r.vaoTriangles->release();
-}
-
-void UiWidgetEngine::destroyTriangles(Int id)
-{
-    TrianglesBuffer& r = trianglesBuffers_[id];
-    r.vaoTriangles->destroy();
-    delete r.vaoTriangles;
-    r.vboTriangles.destroy();
-    trianglesIdGenerator_.release(id);
 }
 
 } // namespace widgets
